requirements.txt
aiobotocore==2.12.2
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
arm_pyart==1.18.0
async-timeout==4.0.3
attrs==23.2.0
botocore==1.34.51
Cartopy==0.22.0
certifi==2024.2.2
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmweather==0.3.2
contourpy==1.2.0
cycler==0.12.1
dask==2024.4.0
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
h5netcdf==1.3.0
h5py==3.10.0
idna==3.6
importlib_metadata==7.1.0
jmespath==1.0.1
kiwisolver==1.4.5
lat-lon-parser==1.3.0
locket==1.0.0
matplotlib==3.8.3
mda-xdrlib==0.2.0
memory-profiler==0.61.0
multidict==6.0.5
netCDF4==1.6.5
numpy==1.26.4
open-radar-data==0.1.0
packaging==24.0
pandas==2.2.1
partd==1.4.1
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
psutil==5.9.8
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3fs==2024.3.1
scipy==1.12.0
shapely==2.0.3
six==1.16.0
toolz==0.12.1
tzdata==2024.1
urllib3==2.0.7
wrapt==1.16.0
xarray==2024.3.0
xarray-datatree==0.0.14
xmltodict==0.13.0
xradar==0.5.0
yarl==1.9.4
